+++ refactored/src/index.js
@@ -1,27 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { createBrowserHistory } from "history";
-import { Router, Route, Switch } from "react-router-dom";
-
+import App from 'components/App.jsx';
 import "assets/css/material-dashboard-react.css?v=1.4.1";
 
-import indexRoutes from "routes/index.jsx";
+const history = createBrowserHistory();
 
-const hist = createBrowserHistory();
-
-ReactDOM.render(
-  <Router history={hist}>
-    <Switch>
-      {indexRoutes.map((prop, key) => {
-        return (
-          <Route
-            path={prop.path}
-            component={prop.component}
-            key={key}
-          />
-        );
-      })}
-    </Switch>
-  </Router>,
-  document.getElementById("root")
-);
+ReactDOM.render(<App history={history} />, document.getElementById("root"));