+++ refactored/src/views/Icons/IconsView.jsx
@@ -12,7 +12,7 @@ import CardBody from "components/Card/CardBody.jsx";
 
 import iconsStyle from "assets/jss/material-dashboard-react/views/iconsStyle.jsx";
 
-function Icons(props) {
+function IconsView(props) {
   const { classes } = props;
   return (
     <GridContainer>
@@ -64,8 +64,8 @@ function Icons(props) {
   );
 }
 
-Icons.propTypes = {
+IconsView.propTypes = {
   classes: PropTypes.object.isRequired
 };
 
-export default withStyles(iconsStyle)(Icons);
+export default withStyles(iconsStyle)(IconsView);