+++ refactored/src/views/Notifications/NotificationsView.jsx
@@ -44,7 +44,7 @@ const styles = {
   }
 };
 
-class Notifications extends React.Component {
+class NotificationsView extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -297,4 +297,4 @@ class Notifications extends React.Component {
   }
 }
 
-export default withStyles(styles)(Notifications);
+export default withStyles(styles)(NotificationsView);