+++ original/CHANGELOG.md
@@ -103,7 +103,7 @@
 
 ## [1.2.0] 2018-04-16
 ### Bug Fixing
-- Renamed `src/containers/App/App.jsx` to `src/layouts/Dashboard/Dashboard.jsx`
+- Renamed `src/containers/App/App.jsx` to `src/layouts/DashboardView/DashboardLayout.jsx`
 - Renamed `src/variables/styles.jsx` to `src/assets/jss/material-dashboard-react.jsx`
 - Moved all the files from `src/variables/styles/` to `src/assets/jss/material-dashboard-react/`
 - Changes caused by the upgrade of `material-ui`
@@ -130,4 +130,4 @@
 ## [1.0.0] 2018-01-29
 ### Original Release
 - Added Material-UI as base framework
-- Added design from Material Dashboard by Creative Tim
+- Added design from Material DashboardView by Creative Tim